+++ datatypes/src/jvmTest/kotlin/PublicKeyTest.kt
@@ -1,24 +1,17 @@
 import at.asitplus.KmmResult.Companion.wrap
 import at.asitplus.crypto.datatypes.CryptoPublicKey
-import at.asitplus.crypto.datatypes.EcCurve
 import at.asitplus.crypto.datatypes.asn1.Asn1Element
 import at.asitplus.crypto.datatypes.asn1.Asn1Sequence
 import at.asitplus.crypto.datatypes.asn1.parse
 import at.asitplus.crypto.datatypes.fromJcaPublicKey
 import at.asitplus.crypto.datatypes.getJcaPublicKey
 import at.asitplus.crypto.datatypes.io.Base64Strict
-import at.asitplus.crypto.datatypes.io.BitSet
-import at.asitplus.crypto.datatypes.io.toBitSet
-import at.asitplus.crypto.datatypes.misc.UVarInt
-import com.ionspin.kotlin.bignum.integer.toBigInteger
 import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
-import io.kotest.matchers.string.shouldStartWith
 import io.matthewnelson.encoding.base16.Base16
-import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
@@ -34,10 +27,6 @@ import java.security.interfaces.RSAPublicKey
 
 class PublicKeyTest : FreeSpec({
     Security.addProvider(BouncyCastleProvider())
-
-
-
-
 
     "EC" - {
         withData(256, 384, 521) { bits ->