+++ CHANGELOG.md
@@ -50,9 +50,19 @@
 #### 2.2.1
 * Update conventions
   * Rename CBOR annotations
+  * Target Java 17
 
 ### 2.3.0
 * Change `CryptoPublicKey.toJsonWebKey()` return type from `KmmResult<JsonWebKey>` to `JsonWebKey`
 * Add `CryptoSignature.parseFromJca` function
 * Refactor `CryptoPublicKey.keyID` to `CryptoPublicKey.didEncoded` to better reflect what it actually is
 * Rename `CryptoPublicKey.fromKeyId` to `CryptoPublicKey.fromDid`
+
+### 2.4.0
+* Add Support for EC Point compression
+* Add Support for full Cose-Key Spec
+* Correct Multibase Encoding
+* Change `DID:KEY` encoding to Base58_BTC to comply with draft
+* Add Multibase Encoder/Decoder
+* Add UVarInt datatype (63 bit max)
+* Remove MultibaseHelper